Use global PouchDB object instead of requiring module (#10)
This makes it easier to use in the browser: just add a script tag with
the hosted pouchdb.min.js. It also makes it easier to use custom
builds of PouchDB, like the pouchdb-node or pouchdb-browser builds.

To be revisited when JavaScript modules are less of a mess. Surprised
as I am, this seems like a problem functors (ML-style, on the module
level) would solve.

Also in this commit:
- test against PouchDB to 6.4.2
- use make instead of pulp for building and running tests